Revert "Plugin admin dashboard redesign"
