--- .svgrrc.json
@@ -1,15 +1,0 @@
-{
-  "svgo": true,
-  "svgoConfig": {
-    "plugins": [
-      {
-        "name": "preset-default",
-        "params": {
-          "overrides": {
-            "removeViewBox": false
-          }
-        }
-      }
-    ]
-  }
-}
