+++ composer.json
@@ -103,7 +103,7 @@
     "newfold-labs/wp-module-installer": "^1.6.5",
     "newfold-labs/wp-module-link-tracker": "^1.1.4",
     "newfold-labs/wp-module-loader": "^1.0.12",
-    "newfold-labs/wp-module-marketplace": "^3.1.0",
+    "newfold-labs/wp-module-marketplace": "^3.0.4",
     "newfold-labs/wp-module-migration": "^1.6.0",
     "newfold-labs/wp-module-next-steps": "^1.3.0",
     "newfold-labs/wp-module-notifications": "^1.6.12",
@@ -115,7 +115,7 @@
     "newfold-labs/wp-module-pls": "^1.1.2",
     "newfold-labs/wp-module-runtime": "^1.1.3",
     "newfold-labs/wp-module-secure-passwords": "^1.1.2",
-    "newfold-labs/wp-module-solutions": "^2.5.0",
+    "newfold-labs/wp-module-solutions": "^2.4.0",
     "newfold-labs/wp-module-sso": "^1.1.1",
     "newfold-labs/wp-module-staging": "^2.5.1",
     "wp-forge/wp-update-handler": "^1.0.2",
