+++ src/app/components/app-nav/index.js
@@ -3,93 +3,63 @@ import apiFetch from '@wordpress/api-fetch';
 import { useViewportMatch } from '@wordpress/compose';
 import { addQueryArgs, cleanForSlug } from '@wordpress/url';
 import { filter } from 'lodash';
-import { Button, AppBarNavigation, useNavigationContext } from '@newfold/ui-component-library';
+import { Modal, SidebarNavigation } from '@newfold/ui-component-library';
 import { default as NewfoldNotifications } from '@modules/wp-module-notifications/assets/js/components/notifications/';
 import { NavLink, useLocation } from 'react-router-dom';
-import classnames from 'classnames';
-import { getPlatformPathUrl } from 'App/util/helpers';
-import { getEditorUrl } from 'App/util/themeUtils';
+import { Bars3Icon } from '@heroicons/react/24/outline';
 import { topRoutes, utilityRoutes } from 'App/data/routes';
-import { RectangleGroupIcon, ArrowUpRightIcon } from '@heroicons/react/24/outline';
-import Logo from '../logo';
+import { handleHelpLinksClick } from '../../util/helpers';
+import Logo from './logo';
 
-export const AppNavHeader = () => {
-	const { mode } = AppBarNavigation.AppBar.useContext();
+export const SideNavHeader = () => {
 	return (
-		<Logo variant={ mode === 'inline' ? 'icon' : 'wordmark' }/>
+		<header className="nfd-pt-2 nfd-px-3 nfd-mb-6 nfd-space-y-6">
+			<Logo />
+		</header>
 	);
 };
 
-export const AppNavMenu = () => {
+export const SideNavMenu = () => {
 	const location = useLocation();
-	const { setActivePath, activePath } = useNavigationContext();
 
-	const menu = () => {
+	const primaryMenu = () => {
 		return (
-			<AppBarNavigation.AppBar.Nav>
-				{ [ ...topRoutes, ...utilityRoutes ]?.map(
-					( page ) => {
-						if ( true !== page.condition ) {
-							return null;
-						}
-
-						const { mode, setOpen } = AppBarNavigation.AppBar.useContext();
-						return (
-							<AppBarNavigation.AppBar.Item
+			<ul className="nfd-flex nfd-flex-col nfd-gap-1.5">
+				{ topRoutes.map(
+					( page ) =>
+						true === page.condition && (
+							<SideNavMenuItem
 								key={ page.name }
 								label={ page.title }
-								href={ `#${ page.name }` }
-								className={ classnames(
-									'group-[.nfd-appbar-item--active]:nfd-text-[var(--color-primary)]',
-									{
-										'nfd-px-8 group-[.nfd-appbar-item--active]:nfd-bg-[#DBF1FC80] group-[.nfd-appbar-item--active]:nfd-font-bold': 'inline' === mode,
-										'nfd-px-0 nfd-font-bold nfd-bg-transparent group-[.nfd-appbar-item]:!nfd-bg-transparent  hover:!nfd-bg-black': 'collapsed' === mode,
-									}
-								) }
-								onClick={ ( e) => {
-									// set open to false
-									setOpen( false );
-									// run any actions assigned to page
-									( !! page.action && page.action instanceof Function ) ? 
-									page.action( e ) :
-									null;
-								} }
+								name={ page.name }
+								icon={ page.Icon }
+								path={ page.name }
+								action={ page.action }
+								subItems={ page.subRoutes }
 							/>
 						)
-					}
 				) }
-			</AppBarNavigation.AppBar.Nav>
+			</ul>
 		);
 	};
 
-	const actions = () => {
+	const secondaryMenu = () => {
 		return (
-			<>
-				<Button
-					as={ 'a' }
-					className={ 'nfd-flex nfd-gap-2 nfd-mr-4' }
-					href={ getEditorUrl( 'edit' ) }
-				>
-					Go to AI Editor
-					<RectangleGroupIcon/>
-				</Button>
-				<Button
-					as={ 'a' }
-					className={ 'nfd-flex nfd-gap-2 nfd-mr-4' }
-					href={ window.NewfoldRuntime.linkTracker.addUtmParams(
-						getPlatformPathUrl(
-							'hosting/details',
-							'app/#/sites'
-						)
-					) }
-					variant={ 'secondary' }
-				>
-					Go to Hosting Panel
-					<ArrowUpRightIcon/>
-				</Button>
-			</>
-		)
-	}
+			<ul className="nfd-flex nfd-flex-col nfd-gap-1.5 nfd-mt-4 nfd-pt-4 nfd-border-t nfd-border-[#D8DEE4]">
+				{ utilityRoutes.map( ( page ) => (
+					<SideNavMenuItem
+						key={ page.name }
+						label={ page.title }
+						name={ page.name }
+						icon={ page.Icon }
+						path={ page.name }
+						action={ page.action }
+						subItems={ page.subRoutes }
+					/>
+				) ) }
+			</ul>
+		);
+	};
 
 	const SubMenusManager = () => {
 		// close any open submenus
@@ -116,48 +86,85 @@ export const AppNavMenu = () => {
 	};
 
 	useEffect( () => {
-		if ( location?.pathname ) {
-			setActivePath( `#${ location.pathname }` );
-		}
 		SubMenusManager();
 		document.onclick = SubMenusManager;
 	}, [ location ] );
 
 	return (
-		<>
-			{ menu() }
-			<div className={ 'nfd-grow' }/>
-			{ actions() }
-		</>
+		<div className="nfd-px-0.5 nfd-space-y-6">
+			{ primaryMenu() }
+			{ secondaryMenu() }
+		</div>
 	);
 };
 
-// export const SideNavMenuSubItem = ( { label, path, action } ) => {
-// 	return (
-// 		<li className="nfd-m-0 nfd-pb-1">
-// 			<NavLink
-// 				onClick={ action && action instanceof Function ? action : null }
-// 				to={ path }
-// 				className={ `wppbh-app-subnavitem wppbh-app-subnavitem-${ cleanForSlug(
-// 					label
-// 				) } nfd-flex nfd-items-center nfd-gap-3 nfd-px-3 nfd-py-2 nfd-rounded-md nfd-text-sm nfd-font-medium nfd-text-body leading-none hover:nfd-bg-slate-50 [&.active]:nfd-bg-[#E2E8F0] [&.active]:nfd-text-title` }
-// 			>
-// 				{ label }
-// 			</NavLink>
-// 		</li>
-// 	);
-// };
+export const SideNavMenuItem = ( {
+	label,
+	icon: Icon = null,
+	path,
+	action,
+	subItems,
+} ) => {
+	return (
+		<li className="nfd-mb-0">
+			<NavLink
+				onClick={ action && action instanceof Function ? action : null }
+				to={ path }
+				className={ `wppbh-app-navitem wppbh-app-navitem-${ cleanForSlug(
+					label
+				) } nfd-flex nfd-items-center nfd-gap-3 nfd-px-3 nfd-py-2 nfd-rounded-md nfd-text-sm nfd-font-medium nfd-text-title leading-none hover:nfd-bg-slate-50 [&.active]:nfd-bg-[#E2E8F0]` }
+			>
+				{ Icon && (
+					<Icon className="nfd-flex-shrink-0 nfd--ml-1 nfd-h-6 nfd-w-6" />
+				) }
+				{ label }
+			</NavLink>
 
-export const AppBarNav = () => {
+			{ subItems && subItems.length > 0 && (
+				<ul className="wppbh-app-navitem-submenu nfd-hidden nfd-ml-8 nfd-m-2 nfd-space-y-1.5">
+					{ subItems.map( ( subItem ) => (
+						<SideNavMenuSubItem
+							key={ subItem.name }
+							label={ subItem.title }
+							name={ subItem.name }
+							path={ subItem.name }
+							action={ subItem.action }
+						/>
+					) ) }
+				</ul>
+			) }
+		</li>
+	);
+};
+
+export const SideNavMenuSubItem = ( { label, path, action } ) => {
+	return (
+		<li className="nfd-m-0 nfd-pb-1">
+			<NavLink
+				onClick={ action && action instanceof Function ? action : null }
+				to={ path }
+				className={ `wppbh-app-subnavitem wppbh-app-subnavitem-${ cleanForSlug(
+					label
+				) } nfd-flex nfd-items-center nfd-gap-3 nfd-px-3 nfd-py-2 nfd-rounded-md nfd-text-sm nfd-font-medium nfd-text-body leading-none hover:nfd-bg-slate-50 [&.active]:nfd-bg-[#E2E8F0] [&.active]:nfd-text-title` }
+			>
+				{ label }
+			</NavLink>
+		</li>
+	);
+};
+
+export const SideNav = () => {
 	const location = useLocation();
 	const hashedPath = '#' + location.pathname;
 
 	return (
-		<>
-			<AppBarNavigation.AppBar position={ 'absolute' } className={ 'nfd-pr-2' } collapseAt={ 'l' }>
-				<AppNavHeader/>
-				<AppNavMenu/>
-			</AppBarNavigation.AppBar>
+		<aside className="wppbh-app-sidenav nfd-shrink-0 nfd-hidden min-[783px]:nfd-block nfd-pb-6 nfd-bottom-0 nfd-w-56">
+			<SidebarNavigation>
+				<SidebarNavigation.Sidebar>
+					<SideNavHeader />
+					<SideNavMenu />
+				</SidebarNavigation.Sidebar>
+			</SidebarNavigation>
 			<NewfoldNotifications
 				constants={ {
 					context: 'bluehost-app-nav',
@@ -171,11 +178,56 @@ export const AppBarNav = () => {
 					useEffect,
 				} }
 			/>
-		</>
+		</aside>
+	);
+};
+
+export const MobileNav = () => {
+	const [ isOpen, setIsOpen ] = useState( false );
+
+	const location = useLocation();
+	// Close mobile nav when location changes
+	useEffect( () => {
+		setIsOpen( false );
+	}, [ location ] );
+
+	return (
+		<header className="nfd-sticky nfd-z-30 nfd-top-0 min-[600px]:nfd-top-[46px] nfd-border-b nfd-border-line">
+			<div className="nfd-flex nfd-justify-between nfd-items-center nfd-bg-white">
+				<div className="nfd-px-4">
+					<Logo />
+				</div>
+				<button
+					id="nfd-app-mobile-nav"
+					className="nfd-h-16 nfd-px-4 nfd-text-body nfd-flex nfd-items-center focus:nfd-outline-none focus:nfd-ring-2 focus:nfd-ring-inset focus:nfd-ring-primary"
+					onClick={ () => {
+						setIsOpen( true );
+					} }
+				>
+					<span className="nfd-sr-only">Open Navingation Menu</span>
+					<Bars3Icon className="nfd-w-6 nfd-h-6" />
+				</button>
+
+				<Modal
+					isOpen={ isOpen }
+					onClose={ () => setIsOpen( false ) }
+					className="wppbh-app-sidenav-mobile nfd-z-40"
+					initialFocus
+				>
+					<Modal.Panel className="nfd-p-0 nfd-overflow-visible">
+						<div className="wppbh-app-sidenav nfd-p-5 nfd-max-h-[70vh] nfd-overflow-y-auto">
+							<SideNavMenu />
+						</div>
+					</Modal.Panel>
+				</Modal>
+			</div>
+		</header>
 	);
 };
 
 export const AppNav = () => {
 	const isLargeViewport = useViewportMatch( 'medium' );
-	return <AppBarNav/>;
+	handleHelpLinksClick();
+
+	return <>{ ( isLargeViewport && <SideNav /> ) || <MobileNav /> }</>;
 };
