+++ src/app/components/icons/BluehostIcon.js
@@ -1,0 +1,18 @@
+const BluehostIcon = ( {} ) => (
+	<svg
+		width="20"
+		height="21"
+		viewBox="0 0 20 21"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+	>
+		<path
+			fillRule="evenodd"
+			clipRule="evenodd"
+			d="M18.8385 6.73946V2H14.1345V6.73946H18.8385ZM18.8385 8.12992V12.8701H14.1345V8.12992H18.8385ZM6.70192 19V14.2598H2V19H6.70192ZM12.7702 14.2598V19H8.06617V14.2598H12.7702ZM18.8385 14.2598V19H14.1345V14.2598H18.8385ZM12.7702 12.8701V8.12992H8.06617V12.8701H12.7702ZM6.70192 8.12992V12.8701H2V8.12992H6.70192ZM12.7667 2V6.73946H8.06617V2H12.7667ZM6.70192 6.73946V2H2V6.73946H6.70192Z"
+			fill="white"
+		/>
+	</svg>
+);
+
+export default BluehostIcon;
