+++ src/app/components/icons/WordPressIcon.js
@@ -1,0 +1,46 @@
+const WordPressIcon = ( {} ) => {
+	return (
+		<svg
+			width="20"
+			height="21"
+			viewBox="0 0 20 21"
+			fill="none"
+			xmlns="http://www.w3.org/2000/svg"
+		>
+			<g clipPath="url(#clip0_4746_6364)">
+				<path
+					d="M1.42148 10.5001C1.42148 13.8958 3.39487 16.8304 6.25644 18.221L2.16421 7.00879C1.68821 8.07572 1.42148 9.25642 1.42148 10.5001Z"
+					fill="#212936"
+				/>
+				<path
+					d="M15.7917 10.0675C15.7917 9.00725 15.4109 8.273 15.0842 7.7015C14.6494 6.99484 14.2418 6.39641 14.2418 5.68975C14.2418 4.90114 14.8399 4.16706 15.6823 4.16706C15.7204 4.16706 15.7565 4.17179 15.7935 4.17392C14.2672 2.77561 12.2338 1.92188 10.0003 1.92188C7.00327 1.92188 4.36647 3.45958 2.83252 5.78867C3.03379 5.79471 3.22348 5.79896 3.38459 5.79896C4.28192 5.79896 5.67091 5.69008 5.67091 5.69008C6.13337 5.66282 6.18789 6.34205 5.72592 6.39674C5.72592 6.39674 5.26118 6.45142 4.74404 6.47852L7.86811 15.7709L9.74551 10.1403L8.40892 6.47819C7.94695 6.45109 7.50931 6.39641 7.50931 6.39641C7.04702 6.36931 7.10121 5.66249 7.5635 5.68975C7.5635 5.68975 8.98025 5.79863 9.82321 5.79863C10.7204 5.79863 12.1095 5.68975 12.1095 5.68975C12.5723 5.66249 12.6267 6.34172 12.1645 6.39641C12.1645 6.39641 11.6988 6.45109 11.1827 6.47819L14.2829 15.7002L15.1386 12.8407C15.5095 11.6542 15.7917 10.8019 15.7917 10.0675Z"
+					fill="#212936"
+				/>
+				<path
+					d="M10.1508 11.25L7.57687 18.7294C8.3454 18.9554 9.15817 19.0789 10.0003 19.0789C10.9993 19.0789 11.9574 18.9062 12.8492 18.5926C12.8261 18.5559 12.8052 18.5169 12.7881 18.4745L10.1508 11.25Z"
+					fill="#212936"
+				/>
+				<path
+					d="M17.5279 6.38379C17.5648 6.65705 17.5857 6.95039 17.5857 7.26593C17.5857 8.13648 17.4231 9.1151 16.9334 10.3387L14.3129 17.9151C16.8634 16.4279 18.5788 13.6647 18.5788 10.4999C18.579 9.00834 18.198 7.60579 17.5279 6.38379Z"
+					fill="#212936"
+				/>
+				<path
+					d="M10.0003 0.5C4.48629 0.5 0 4.98596 0 10.5C0 16.0147 4.48629 20.5005 10.0003 20.5005C15.5142 20.5005 20.0011 16.0147 20.0011 10.5C20.001 4.98596 15.5142 0.5 10.0003 0.5ZM10.0003 20.0421C4.73914 20.0421 0.458537 15.7615 0.458537 10.5C0.458537 5.23882 4.73898 0.958537 10.0003 0.958537C15.2613 0.958537 19.5416 5.23882 19.5416 10.5C19.5416 15.7615 15.2613 20.0421 10.0003 20.0421Z"
+					fill="#212936"
+				/>
+			</g>
+			<defs>
+				<clipPath id="clip0_4746_6364">
+					<rect
+						width="20"
+						height="20.0005"
+						fill="white"
+						transform="translate(0 0.5)"
+					/>
+				</clipPath>
+			</defs>
+		</svg>
+	);
+};
+
+export default WordPressIcon;
