+++ src/app/components/icons/index.js
@@ -1,10 +1,2 @@
-import { ReactComponent as BluehostIcon } from './BluehostIcon.svg';
-
-export { BluehostIcon };
-export { ReactComponent as WordPressIcon } from './WordPressIcon.svg';
-export { ReactComponent as PartyIcon } from './PartyIcon.svg';
-export { ReactComponent as AiEditorIcon } from './AiEditorIcon.svg';
-export { ReactComponent as YoastLogoIcon } from './YoastLogoIcon.svg';
-export { ReactComponent as HelpIcon } from './HelpIcon.svg';
-
-console.log(BluehostIcon)
+export { default as BluehostIcon } from './BluehostIcon';
+export { default as WordPressIcon } from './WordPressIcon';
