+++ src/app/data/routes.js
@@ -11,7 +11,6 @@ import { Route, Routes } from 'react-router-dom';
 import Home from '../pages/home';
 import Store from '../pages/ecommerce/page';
 import Marketplace from '../pages/marketplace';
-import Commerce from '../pages/commerce';
 import Settings from '../pages/settings';
 import Help from '../pages/help';
 import Admin from '../pages/admin';
@@ -102,7 +101,6 @@ export const AppRoutes = () => {
 			<Route path="/staging" element={ <StagingRedirect /> } />
 			<Route path="/performance" element={ <PerformanceRedirect /> } />
 			<Route path="/hosting" element={ <HostingRedirect /> } />
-			<Route path="/staging" element={ <StagingRedirect /> } />
 			<Route path="/" element={ <Home /> } />
 			<Route
 				path="*"
@@ -121,7 +119,7 @@ export const AppRoutes = () => {
 	);
 };
 
-const topRoutePaths = [ '/home', '/commerce', '/marketplace', '/settings' ];
+const topRoutePaths = [ '/home', '/store', '/marketplace', '/settings' ];
 const utilityRoutePaths = [ '/help' ];
 
 export const routes = [
@@ -170,12 +168,6 @@ export const routes = [
 		condition: true,
 	},
 	{
-		name: '/commerce',
-		title: __( 'Commerce', 'wp-plugin-bluehost' ),
-		Component: Commerce,
-		condition: true,
-	},
-	{
 		name: '/marketplace',
 		title: __( 'Marketplace', 'wp-plugin-bluehost' ),
 		Component: Marketplace,
@@ -191,41 +183,16 @@ export const routes = [
 		condition: true,
 	},
 	{
-		name: '/settings/settings',
-		title: __( 'Settings', 'wp-plugin-bluehost' ),
-		Component: Settings,
-		Icon: AdjustmentsHorizontalIcon,
-		condition: true,
-	},
-	{
-		name: '/settings/staging',
-		title: __( 'Staging', 'wp-plugin-bluehost' ),
-		Component: Settings,
-		Icon: AdjustmentsHorizontalIcon,
-		condition: true,
-	},
-	{
-		name: '/settings/performance',
-		title: __( 'Performance', 'wp-plugin-bluehost' ),
-		Component: Settings,
-		Icon: AdjustmentsHorizontalIcon,
-		condition: true,
-	},
-	{
 		name: '/help',
-		title: __( 'Help', 'wp-plugin-bluehost' ),
+		title: __( 'Help with WordPress', 'wp-plugin-bluehost' ),
 		Component: Help,
 		Icon: HelpIcon,
 		condition: true,
-		action: function ( e ) {
-			if (
-				NewfoldRuntime.hasCapability( 'canAccessHelpCenter' ) &&
-				window.NewfoldFeatures.isEnabled( 'helpCenter' )
-			) {
-				e.preventDefault();
-				window.newfoldEmbeddedHelp.toggleNFDLaunchedEmbeddedHelp();
-			}
-		},
+		action:
+			NewfoldRuntime.hasCapability( 'canAccessHelpCenter' ) &&
+			( await window.NewfoldFeatures.isEnabled( 'helpCenter' ) )
+				? HelpCenterAI
+				: false,
 	},
 	{
 		name: '/admin',
