+++ src/app/index.js
@@ -14,11 +14,10 @@ import { ErrorBoundary } from 'react-error-boundary';
 import { store as noticesStore } from '@wordpress/notices';
 import { kebabCase, filter } from 'lodash';
 import { useHandlePageLoad } from './util/hooks';
-import { Root, NavigationProvider } from '@newfold/ui-component-library';
+import { Root } from '@newfold/ui-component-library';
 import { NotificationFeed } from './components/notifications';
 import { handleHelpLinksClick } from './util/helpers';
-import Logo from './components/logo';
-import { AppNav } from './components/app-nav/index';
+import Logo from './components/app-nav/logo';
 
 // component sourced from module
 import { default as NewfoldNotifications } from '@modules/wp-module-notifications/assets/js/components/notifications/';
@@ -142,10 +141,10 @@ const AppBody = ( props ) => {
 				} }
 			/>
 			<div className="wppbh-app-body">
-				<header className="nfd-mb-20">
-					<AppNav />
+				<header className="nfd-mb-6">
+					<Logo />
 				</header>
-				<div className="wppbh-app-body-inner nfd-flex nfd-justify-center">
+				<div className="wppbh-app-body-inner">
 					<ErrorBoundary FallbackComponent={ <ErrorCard /> }>
 						{ hasError && <ErrorCard error={ hasError } /> }
 						{ ( true === booted && <AppRoutes /> ) ||
@@ -166,11 +165,9 @@ export const App = () => (
 		<Root context={ { isRtl: false } }>
 			<NotificationFeed>
 				<Router>
-					<NavigationProvider>
-						<div className="wppbh-app-container min-[783px]:nfd-p-8 min-[783px]:nfd-flex nfd-gap-6 nfd-max-w-full xl:nfd-max-w-screen-xl 2xl:nfd-max-w-screen-2xl nfd-my-0 nfd-mx-auto">
-							<AppBody />
-						</div>
-					</NavigationProvider>
+					<div className="wppbh-app-container min-[783px]:nfd-p-8 min-[783px]:nfd-flex nfd-gap-6 nfd-max-w-full xl:nfd-max-w-screen-xl 2xl:nfd-max-w-screen-2xl nfd-my-0">
+						<AppBody />
+					</div>
 				</Router>
 			</NotificationFeed>
 		</Root>
