+++ src/app/pages/home/SolutionCard.js
