--- src/app/pages/home/cards/index.js
@@ -1,17 +1,0 @@
-import { default as SolutionCard } from './SolutionCard';
-import { default as ExpertCard } from './ExpertCard';
-import { default as HelpCard } from './HelpCard';
-import { default as ProDesignCard } from './ProDesignCard';
-import { default as ReferralProgramCard } from './ReferralProgramCard';
-import { default as QuickLinksCard } from './QuickLinksCard';
-
-export default () => (
-	<>
-		<SolutionCard />
-		<ExpertCard />
-		<HelpCard />
-		<ProDesignCard />
-		<ReferralProgramCard />
-		<QuickLinksCard />
-	</>
-)
