--- src/app/pages/home/quick-links/index.js
@@ -1,16 +1,0 @@
-import { default as YoastQuickLink } from './yoast';
-import { default as SettingsQuickLink } from './settings';
-import { default as AiBuilderQuickLink } from './ai-builder';
-import { default as AiEditorQuickLink } from './ai-editor';
-import { default as HelpQuickLink } from './help';
-
-
-export default () => (
-	<>
-		<YoastQuickLink />
-		<SettingsQuickLink />
-		<AiBuilderQuickLink />
-		<AiEditorQuickLink />
-		<HelpQuickLink />
-	</>
-)
