--- src/app/pages/home/suggested-videos/index.js
@@ -1,9 +1,0 @@
-import SuggestedVideo from './suggested-video';
-
-export default () => (
-	<>
-		<SuggestedVideo />
-		<SuggestedVideo />
-		<SuggestedVideo />
-	</>
-);
