+++ src/app/pages/marketplace/index.js
@@ -7,24 +7,6 @@ import { NewfoldRuntime } from '@newfold/wp-module-runtime';
 import { default as NewfoldMarketplace } from '@modules/wp-module-marketplace/components/';
 
 const MarketplacePage = () => {
-
-	useEffect( () => {
-		// run when mounts
-		const marketplacePortal = document.getElementById( 'marketplace-portal' );
-		
-		if ( marketplacePortal ) {
-			window.NFDPortalRegistry.registerPortal(
-				'marketplace',
-				marketplacePortal
-			);
-		}
-
-		// run when unmounts
-		return () => {
-			window.NFDPortalRegistry.unregisterPortal( 'marketplace' );
-		};
-	}, [] );
-
 	// constants to pass to module
 	const moduleConstants = {
 		supportsCTB: true,
@@ -72,7 +54,10 @@ const MarketplacePage = () => {
 
 	return (
 		<Page className={ 'wppbh-app-marketplace-page' }>
-			<NewfoldMarketplace constants={ moduleConstants } methods={ moduleMethods } />
+			<NewfoldMarketplace
+				methods={ moduleMethods }
+				constants={ moduleConstants }
+			/>
 		</Page>
 	);
 };
